Refactor stationary energy chat constants
